Do not report all after objects without -fPIC as having lost PIC (#153)
This took a little longer to investigate than I thought.  It also
uncovered some problems with the general algorithm for checking .a
libraries and the .o members in each.  In issue #153, the problem is
explained that not all .o objects in a .a file will have been built
with -fPIC.  This is true, you can have a mix of both.  rpminspect was
incorrectly reporting all objects without -fPIC in the after build as
having lost -fPIC.  That's not true as it assumes the objects had
-fPIC in the before build.

The reporting now shows objects that lost -fPIC between the before and
after build but also lists any objects in the after build without
-fPIC as just not having it.  A loss of -fPIC is reported as BAD while
reporting missing -fPIC is just INFO because it may be deliberate.