Drop eu-elfcmp(1) usage in the changedfiles inspection.
Using eu-elfcmp to see if ELF libraries have changed is not very
useful in the context of rpminspect.  We have annocheck and a coming
abicheck inspection that will provide better analysis.  Discontinue
using eu-elfcmp in changedfiles as well as skipping ELF files from the
changedfiles inspection.  This should cut down on the noise from
rpminspect for non-issues.