Rename on_stat_whitelist() to on_stat_whitelist_mode(), fix some errors.
This function was only checking the mode_t for files, so rename it
accordingly.  It's used by both the permissions and addedfiles
inspections.  Make sure it returns true only when the file is found on
the whitelist, false otherwise.  And make the final reporting only
occur if the file is not on the whitelist and has setuid or setgid
set.