Add on_stat_whitelist_owner() and on_stat_whitelist_group()
Hook up these functions in the 'ownership' inspection as well.  If
invalid owners and groups are found, check the whitelist to see if
it's deliberate.  If it's on the whitelist, report it as INFO.  If
it's on the whitelist but does not match, report as VERIFY.  Otherwise
report according to the regular code path.