Expand foreach_peer_file() with use_ignores parameter.
This is to implement the 'ignore' section of the rpminspect
configuration file.  For certain inspections we want rpminspect to
ignore paths specified using glob(3) syntax in the configuration file.
globs can be in the main configuration file or be in the profile or be
in the local-directory config file.

Modify all the calls to foreach_peer_file() to pass true to honor the
ignores and false to ignore them.  The inspections where ignores
are..ignored are in the addedfiles and removedfiles inspections as
well as inspections that deal primarily with source RPMs or the
metadata.  The main purpose of ignores is to tell rpminspect to skip
content validation and/or comparisons for certain types of files.  A
common one might be *.pyo and *.pyc files.