Support a positional parameter on 'make check' to run part of test suite.
This is useful for running only part of the test suite, which is
common when debugging one particular inspection.  Let's say I only
want to run test_elf.py over and over.  I can do that now with:

    make check elf

This avoids having to set the environment variables and run the python
command as noted in the README.  Just makes it simpler.