'sort | uniq' -> 'sort -u'
Old habits die hard.  Whenever I write scripts that use this
construct, I still think to myself "I might want to run this on a
SunOS 4 system" -or- "what if I run this on a SunOS 5 system?  can't
assume /usr/xpg4/bin is in the PATH!".  And then I think, "huh, wait,
why do I care about that anymore?".  It's 2020 and -u is defined in
POSIX for sort(1), so I'm going to call that good enough.