Rename 'make release' to 'make new-release'; add 'make release'
'make new-release' does a version bump, tag, and push.  It runs
./utils/release.sh with the -A option.  'make release' just does a tag
and push, so use this if you do not want to make a new version number
(it runs ./utils/release.sh -t -p).  The 'make release' target is
intended for making new official releases where the version number has
already been incremented as part of the development process.