Support multiple configuration files.
This change is notable because it changes the behavior of rpminspect
when run.  Previously, rpminspect looked for
/etc/rpminspect/rpminspect.yaml and if it couldn't find it, asked you
to use -c to specify a config file.  Now rpminspect does the
following:

    * If you do not specify a config file with -c:
          * Try /usr/share/rpminspect/rpminspect.yaml, if found then read it in
          * If no option exists, exit out and ask the user for a -c option
    * If you specify a config file with -c:
          * Try the string provided, if found then read it in
          * If the file does not exist or is unreadable, error out

In all cases, rpminspect also looks for a ./rpminspect.yaml where you
launch rpminspect.  If that's found, it reads it in and overrides the
configuration already read in.

With this behavior, it is possible to only run rpminspect with a local
rpminspect.yaml file.  The data packages have changed as a result to
move all files in to /usr/share/rpminspect.  No data package will
provide rpminspect.yaml.  In rpminspect-data-generic, there is
/usr/share/rpminspect/generic.yaml

The user is expected to specify the configuration file is relying on a
data package.  Data packages are encouraged to provide a launcher
script to make this easier.

The RPM 'Requires' of rpminspect-data has been dropped as well.  This
is no longer a requirement, but just a nice to have.  The new
structure should allow concurrent installation of all of the data
packages.

As a result of this change, increment the version number to 1.1.