Add the '%files' inspection to librpminspect
The was a remaining thing to port from rpminspect's ancestor.  The
%files inspection looks at the %files blocks in the spec file and
raises a VERIFY result if any forbidden path references are found.
Forbidden references are defined in the configuration file in the
'files' section.  The generic.yaml configuration file has:

    files:
        forbidden_paths:
            - /usr/lib

The purpose of the inspection is to ensure package spec files use path
macros as much as possible.  Rather than /usr/lib, packages should
rely on %{_libdir}.  Vendors should make sure this list covers all
explicit path references they care about.

This this setting is undefined in the configuration file, the
inspection will pass.  The inspection also passes when there is no
source RPM available, but produces an INFO message saying this
inspection is only valid for source RPMs.