Added the 'types' inspection to compare MIME types between builds.
This inspection compares MIME types of files between builds and
reports changes.  For example, it /usr/bin/program is
"application/x-pie-executable" in the before build and becomes
"text/plain" in the after build, it reports that as a VERIFY result.
This inspection will probably require some adjustment over time, so
keep that in mind.

rpminspect's ancestor compared the human readable strings from file(1)
rather than the MIME types which caused many problems.  Using MIME
types is an attempt by me to simplify the inspection and reduce the
maintenance burden.