Use the 'make instreqs' target for install test suite deps.
Migrate the centos7, centos8, and fedora CI yaml files to using 'make
instreqs' to install the build and test suite dependencies.  The
reason for this is so the Makefile target can be used locally and not
restricted to the GitHub Actions environment.

This adds a block to the top level Makefile to determine the OS we are
on and what packages need to be installed.  Remove the commented out
block of test suite deps from rpminspect.spec.in.

Add utils/determine-os.sh that prints out short strings indicating the
OS we are on, which is used by the Makefile to set the list of
packages to install and the package installation commands.