Move REQS and PIP lists out of the Makefile to files in osdep/
Divide it up by operating system.  Put the system packages in a file
called reqs.txt and put Python modules installed by 'pip' in the
pip.txt file.  Blank lines and lines beginning with '#' are ignored.

The Fedora-derived distributions begin with a list of packages
obtained from the spec file.  That is, all of the BuildRequires and
Requires lines.