Make the reqs.txt files in osdep/ contain all deps
Having the Fedora-derived systems start with a list containing all of
the BuildRequires and Requires from the spec file template is
confusing.  It also skipped macro processing which is a problem since
some of the BuildRequires are OS-dependent.

Move to having the reqs.txt files contain all of the packages for each
OS.  Separate spec files will be needed anyway for different RPM-based
distributions, so there's no real good reason to use it for the 'make
instreqs' target.