Adjust test_metadata.LosingVendorCompareKojiBuild to expect VERIFY
This test was not being run prior to #165 as there was a second test
class shadowing it. Based on the comment on the test, I believe it  was
intended to expect a "VERIFY" result rather than the "OK" result in the
test itself (which is failing). Adjust the test to match the documented
test case.

Fixes: cf5f19b89633 ("python: rename several duplicate test cases")