Improve the error reporting for test result checking
While developing a new test, I provided the wrong waiver_auth and got a
test that failed with:

Traceback (most recent call last):
  File "/home/jcline/devel/rpminspect/test/baseclass.py", line 419, in
  runTest
      self.assertTrue(
      AssertionError: False is not true

It is indeed true that False is not true, but I had to do a bit of
digging to determine what was wrong with my test.

This changes check_results to raise an AssertionError with details about
what exactly was wrong with the results. For example, the issue that generated
the above error now generates:

Traceback (most recent call last):
  File "/home/jcline/devel/rpminspect/test/baseclass.py", line 414, in runTest
    check_results(self.results, self.label, self.result, self.waiver_auth)
  File "/home/jcline/devel/rpminspect/test/baseclass.py", line 67, in check_results
    raise AssertionError(
AssertionError: Expected result=VERIFY with waiver authorization=Not Waivable in [
    {
        "message": "/some/file grew by +2000% on x86_64",
        "result": "VERIFY",
        "waiver authorization": "Anyone"
    }
]