Introduce the 'movedfiles' inspection and a lot of other fixes (#155)
This is not my favorite kind of commit, but it's sort of unavoidable.
The addedfiles, removedfiles, changedfiles, and movedfiles inspections
are all inter-related.  This commit introduces the 'movedfiles'
inspection for the first time.  It will currently report files that
move between subpackages from the before to the after build.  I have
more work to do here to get the remaining move checks in place, but I
wanted to commit this now.

The other big changes:

1) For addedfiles, removedfiles, and movedfiles, if the you are
comparing a rebased build then the reporting level is INFO.  If you
are not comparing rebased builds, then the reporting level starts at
VERIFY.  What is a rebased build?  It's one where the package names
are the same but the version is different.  A non-rebased build is one
where both the name and version are the same.  This is done because
lots of changes are usually expected in rebased builds, but when doing
maintenance work we want to be more strict.

2) The is_rebase() function is introduced here.  I'm already working
on refining this because I don't totally like it.  I want to favor
source RPMs, but fall back on binary packages.  This is just internal
implementation stuff.

3) The movedfiles inspection runs before addedfiles and removedfiles
in the main ordering.  This felt logical to me from a reporting
standpoint, but I am open to changing things around here.

There is more coming on this tomorrow, such as test cases and other
improvements.