tests: optionally check the result message
It can be helpful to make a few assertions about the content of the
message. This lets test authors set the expected message. Although it's
not perfect (checking, for example, the message contains a sub-string
isn't possible with this interface) it's a small expansion of the
current test interface and completely optional.