Enable 'permissions' inspect for single build analysis.
There's no reason to restrict the permissions inspection to
comparisons only.  It can report security permission risks in single
build analysis runs, which is what this patch does.

This patch also reports world-writable files and directories with or
without the sticky bit set as a BAD result requiring a Security
waiver.  Packages needing files with these permissions need to make
sure they are listed in the appropriate product release 'fileinfo'
list in the corresponding rpminspect-data package.