Relicense librpminspect (lib/ and include/) as LGPL-3.0-or-later
This commit changes the license on librpminspect from GPL-3.0-or-later
to LGPL-3.0-or-later.  The rpminspect program (src/) is still
GPL-3.0-or-later.

There are 5 source files in lib/ that carry Apache 2.0 licenses
(checksums.c, copyfile.c, mkdirp.c, rmtree.c, and unpack.c), but these
are ok with GPL and LGPL 3.0 or later projects.  The entirety of
librpminspect is still under LGPL-3.0-or-later.