Add new setting to abidiff section of the config file
The 'security_level_threshold' specifies the ABI compat level where
abidiff(1) findings for packages or DSOs at or below this level
trigger a security severity with the results reporting.  The default
is '2' which means abidiff(1) changes found for packages in ABI compat
levels 1 or 2 will report with a security severity.