Add -n/--no-rebase command line option to disable rebase detection
rpminspect by default will try to determine if the two builds you are
comparing constitute a rebase.  A rebase is defined as having matching
package names but different versions.  If the package names and
versions match, then it is not considered a rebase (though the release
numbers can vary...and probably should).  If a run is considered a
rebase, it affects the results reporting.  Things that would be raised
as concerns for a non-rebase run (that is, a maintenance update to a
package) may not be raised at all in a rebase or may be reported at
the INFO level.  You can now disable this functionality entirely by
passing the -n option to the rpminspect command.