Avoid comparing elf files that are not shared libraries
Sometimes, when comparing two packages the abidiff inspection can
wrongly attempt to compare ELF binaries that are not shared
libraries.  For instance, it can try to compare debuginfo files --
which are technically ELF files, but that we should obviously not
compare.  As a result, it sometimes emits spurious errors like this:

	$ rpminspect -k -v  -T abidiff ./workdir/tmp/alsa-lib-1.2.2-2.fc33 ./workdir/tmp/alsa-lib-1.2.3.2-5.fc33
	Running abidiff inspection...        FAIL

	[...]

	abidiff:
	--------
	1) Comparing from /usr/lib/debug/.dwz/alsa-lib-1.2.2-2.fc33.i386 to
	/usr/lib/debug/.dwz/alsa-lib-1.2.3.2-5.fc33.i386 in package alsa-lib-debuginfo on
	i686 revealed ABI differences.
	Result: VERIFY

	[...]

	Details:
	abidiff: failed to read input file ./workdir/local.JV0zj7/before/i686/alsa-lib-debuginfo-1.2.2-2.fc33.i686//usr/lib/debug/.dwz/alsa-lib-1.2.2-2.fc33.i386
	abidiff: could not find the debug info

	[...]

As the abidiff inspection is really meant to compare shared object
files to detect possibly incompatible ABI changes in them, I think we
should restrict it to just those files.

This patch makes sure to only compare ELF files of ET_DYN kind.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>
Signed-off-by: Dodji Seketeli <dodji@seketeli.org>