Trim worksubdir from paths in reported abidiff and kmidiff commands
The full paths include temporary directories generated while running
rpminspect.  The point of reporting the command is to show how the
command was run and not give an explicit command that you can copy and
paste.  By trimming the worksubdir, we can eliminate the tmp directory
noise and still show example paths that indicate if the file was in
the before or after build.