Use is_rebase() in the 'upstream' inspection
Use the is_rebase() function in the 'upstream' inspection to determine
if two packages are a rebase or not.  This was happening, but was done
manually.  The is_rebase() came along after the upstream inspection
was written, so this just needed to update.

There is still a loop in here to detect whether or not you have source
packages in the comparison.  If not, the inspection is skipped since
it can only work if you have source packages included.