Add get_rpmtag_fileflags() to files.c and call from extract_rpm()
The get_rpmtag_fileflags() function takes the RPM header of the
package we are currently extracting and the file index and extracts
the RPMTAG_FILEFLAGS value for that file from the header.
RPMTAG_FILEFLAGS is an array and entries are found with the file's
index number.  If there are no flags for the file in question, you get
0 back.  Store the flags in the 'flags' member of the rpmfile_entry_t
struct.

File flags are used to determine what has been marked with special
macros at build time.  Things like %doc, %config, %license, and
%readme are all recorded in the header as RPMTAG_FILEFLAGS.
librpminspect will use file flags for certain inspections.