Use correct Version and Release values in download_build()
The string used to construct the download URL pulled the Version and
Release from the specific RPM currently being downloaded, but this is
not correct.  The Version and Release comes from the build, just like
the name.  Subpackages in a build can override the Version and
Release, though this is typically not done.