Add the 'config' inspection to librpminspect
The 'config' inspection compares files across builds that are marked
with %config in the spec file.  The objective is to find configuration
file changes and report those for verification.  If the builds
compared are considered a rebase, rpminspect will report findings at
the INFO level.  If they are not a rebase, findings are reported at
the VERIFY level.

In the case of %config entries that are symlinks, librpminspect will
attempt to read the link destination for comparison, but this will
only work for relative symlinks (which, honestly, is what packages
should have).  For symlink %config entries, librpminspect will report
a %config entry changing to/from a symlink as well as if the link
destinations have changed.

Oh, and for contents comparison, whitespace changes are ignored.  So
if a config file reindents things or adds some blank lines, those will
not cause this inspection to fail.