Add the 'doc' inspection to librpminspect
In rpminspect's ancestor, this was part of what was the TEST_CONFIG
test.  I decided to break it out in to a separate inspection from the
'config' inspection.  What it does is look for changes in %doc entries
and reports:

* If the content between %doc files changed (INFO reporting only)

* If a %doc file was lost in the after build (INFO for rebased builds,
  VERIFY otherwise)

* If a new %doc file appears in the after build (INFO for rebased
  builds, VERIFY otherwise)

The main reason for this check is to make sure package updates include
all the docs they want.  Sometimes upstream projects rename
documentation files or move them and it's easy to lose them during an
update.  For example, version 1.15.1 of a program has a README file
but starting with version 1.15.2 it's now called README.md.  The
package maintainer needs to update the %doc listing in the spec file
to make sure the README content is still packaged up.