Rename the '%files' inspection to 'files' (#194)
The names of inspections are listed in the inspections section of the
configuration files and profile files.  The inspection names are used
as tokens in the file with the value being 'on' or 'off'.  With the
'%' in the name of %files, it presented a libyaml YAML_NO_TOKEN error
when parsing.  Rename the inspection to avoid this problem.