Modify baseclass.py to allow 'before' and 'after' NVR tuples
It's common for test cases to need to override the name, version, or
release of a test package.  Keep the defaults in baseclass.py, but let
individual test cases override that by passing a before and/or an
after tuple to the parent class.  The tuple is (name, version,
release) and they are all strings.  If you do not supply a before or
after tuple to the parent class, the existing default is used.