Add a -D/--dump-config option to rpminspect(1)
Previously when you would enable debugging mode with -d/--debug,
rpminspect would first dump the configuration values currently in use
and then proceed with the regular stdout, stderr, and debugging
messages.  This felt a bit cluttered to me because I don't always want
to see the configuration values.

This commit adds the -D/--dump-config option to rpminspect(1) for
doing just that.  You can use it independently of the -d/--debug
option too.