Fix reporting errors in the 'doc' inspection
Similar to 'config', use the global reported variable to suppress OK
results if the driver ever reports an INFO result.  Make sure we look
for the RPMFILE_DOC bit and reporting levels and waiver authorizations
are consistent with the rest of the library.  Discontinue use of the
minor variable and just inline it with the single xasprintf() using
it.