Define a new GitHub Action using utils/gate.sh
The new GitHub Action is called 'gate' and is divided in to OS jobs.
Right now there's only a job for Fedora, but I might add in CentOS or
other versions of Fedora.

The job runs 'make gate' which runs the utils/gate.sh script.  The
objective here is to ensure that rpminspect continues to run fine for
a sampling of packages from Fedora.  The following packages are tested
with rpminspect:

    zsh
    kernel
    python3
    mutt
    emacs
    tmux
    elfutils

The gate job fails if rpminspect exits on any of these test runs with
anything other than 0 or 1 -or- the program segfaults.  The gate job
does not care what rpminspect finds, just that it runs to completion
when comparing packages.