Adjust what things run during with GitHub Actions
There are now 4 actions:

    CI
    Style
    Extra CI
    Gate

The CI and Style actions run on push events and pull requests.  CI has
been reduced to Fedora rawhide and the Fedora latest release.  Style
is the same as before.

The Extra CI action are all of the other OSes that used to be in CI.
This action runs on push events only, so not on pull requests.

The Gate action also runs on push events only instead of also on pull
requests.