Add get_rpm_header_string_array() to librpminspect
Given an RPM header and an rpmTagVal (e.g., RPMTAG_SOURCE), retrieve
the elements of that string array and return them as a string_list_t.
Caller must free the string_list_t.

The function is named this way because you should only use it for RPM
header tags that are string arrays.