Add filecmp() and use that in place of zcmp/bzcmp/xzcmp
The changedfiles inspection would perform comparisons on any
compressed files found in packages, but it would use zcmp, bzcmp, or
xzcmp depending on how the files were compressed.  The idea is that we
don't care about the compression characteristics so long as the
uncompressed content doesn't change.

The basic idea for this comparison came from rpminspect's ancestor, so
it used zcmp, bzcmp, and xzcmp.  Forking and execing external programs
is something librpminspect is trying to minimize so this commit adds
in two functions:

    read_file_bytes()
        Does an mmap() on a file and returns the contents and length.
    filecmp()
        Compares two files.

Using the previously added uncompress_file() function, we can now
perform these comparisons internally and not need external programs.
The decompression is handled by libarchive, so any formats it supports
can be supported by this inspections.  Right now it's limited to gz,
bz2, and xz.