Better explanation as to why the EmptyLicenseTag tests are skipped.
It is not due to a limitation in rpmfluff.  rpmbuild requires the
following tags to appear in a spec file:

    Name
    Version
    Release
    Summary
    License

Excluding the License tag causes rpmbuild to fail.  You cannot specify
it without any data either as rpmbuild will fail in that case too.

At one point rpmbuild did not enforce this and it may change again, so
we carry a check in rpminspect as well and will leave these test cases
here in case rpmbuild allows excluding the License tag in the future.