Support single package URLs for before and after builds (#190)
rpminspect can work on single RPMs or entire Koji builds.  For the
former, rpminspect expected the RPMs to be locally available.  But
there are times where you may want to compare against an RPM already
built and available for download somewhere.  This patch lets you
specify URLs to single RPM packages.  rpminspect will detect this and
download the package and perform the inspections as if you were
providing a local RPM.