Support slightly older versions of libclamav in inspect_virus.c
libclamav version 0.100 or thereabouts do not have a struct
cl_scan_options definition.  Instead the options are just a bitfield
that's an unsigned int.  In these older versions of libclamav there is
a macro called CL_SCAN_STDOPT.  So in the presence of that macro,
assume we don't have struct cl_scan_options and operate with the
previous API.  Otherwise use the newer API.  This gets the program
working on OpenSUSE Leap 15.2 which has clamav 0.100.