Add the 'politics' inspection to librpminspect.
The politics inspection checks for possible politically sensitive
files in packages.  The inspection performs two passes across files
when provided with rules from the rpminspect-data package.

First, rules are defined in per-product files.  Those live in
/usr/share/rpminspect/rpminspect/PRODUCT_RELEASE

A rules file consists of entries per line and each line three columns.
Those columns are:

    - A glob(7) pattern of the filename to match

    - An MD5, SHA-1, or SHA-256 message digest of the file; or '*' for
      all files

    - The keyword 'allow' or 'deny'

The first pass across files looks at the wildcard lines.  These lines
are intended to apply a blanket allow or deny across files that match
that pattern.  The second pass checks rules with specific message
digests against the file we just matched.  The intent here is the
rules would define exceptions to a blanket wild card rule this way.

Files found to match a rule but are allowed are reported at the INFO
level.  Files found to match a rule but are denied are reported at the
BAD level.  Both types of reporting are marked as Not Waivable.