Define DEFAULT_MESSAGE_DIGEST in constants.h and use that.
Internally librpminspect is computing message digests for files and
caching them.  The library is using SHA256SUM right now, but that
could change in the future when we all find out that SHA-256 is
horribly vulnerable to something for some reason.  I made the default
be a constant so I can make that change in one place when the time
comes.  It won't work that way, but I can pretend it will.