Fix RPMFILE_FLAGS handling for %config files (#221)
A number of problems here.  Make sure we pick up the file flags
correctly when unpacking the payload.  Next, fix the 'config'
inspection so it performs the comparision only when the before and
after file carry RPMFILE_CONFIG flags, not when their flag check
values are equal.  Without this, it was scanning every file in the
package.  Add in some more protections to skip over non-regular files
and non-symlinks.