Honor the -a command line option for downloads as well as runtime (#233)
rpminspect has had the -a option to restrict the architectures
examined during a run.  However, that was limited to the peer
detection code and not the download code.  rpminspect would still
always download every architecture and then later make the decision as
to whether or not to use it.  This patch extends the -a handling to
the download side so it covers downloading (or copying in the case of
locally cached builds) and peer detection.