Prevent invalid pointer dereferencing in invalid result in 'patches' (#245)
The patches inspection was not correctly noting when all of the
results were reported so the fall-through 'OK' result was output as
the last result.  However it would hand an invalid pointer in the
params struct to libjson-c and that caused garbage in the output.

This patch prevents that bogus OK message but also ensures any
reporting of that OK message in the future avoids garbage strings.