Avoid reusing the same abi_pkg_entry_t struct in read_abi()
The read loop that creates the abi_list_t struct was not creating new
abi_pkg_entry_t structs for each package.  It was just using the same
one over and over.  That's because the end of the loop needs to reset
the pointer to NULL and continue.  This fixes up the final memory leak
shown in free_abi()