Allow a set of excluded path prefixes in 'pathmigration'
Expand the 'pathmigration' inspection to support path exclusions.  For
example, the kernel package may deliver modules in /lib/modules.  We
may not want rpminspect to warn about modules needing to be in
/usr/lib even if we have set up the rules to warn about libraries in
/lib.  The path exclusions are flexible enough to handle those cases.

The pathmigration section of the configuration file can now have a
block called 'excluded_paths' which lists the excluded path prefixes.
All paths listed must end with a '/'.  rpminspect will add one if it
does not see one at the end of the prefix.  This is to avoid something
like "/lib/modules" matching
"/lib/modulessddfgwerqwefwercver/oopsfile" and ignoring it.