Small change to the way Koji builds are submitted
My thought process is that I will perform a 'make release' which does
the test suite run and then the tag.  It pushes that to GitHub and it
also makes the tarball and signs it.

For a separate step I will run 'make koji' and tell it what dist-git
branches to build for with the BRANCHES environment variable.  This
step will run 'make srpm' to get an updated spec file.  It uses the
new spec file and the tarball and signature generated from 'make
release' to do the builds in dist-git.

I would like signature verification to happen, but I need to
conditionalize that out for Copr builds.