[lib] Require libabigail >= 1.8 in rpminspect.spec.in
Starting with libabigail 1.8 the abidiff program accepts multiple
header-dir arguments.  This was a requested upstream addition for
abidiff that makes it easier to run from rpminspect.  We need to make
sure we are running with that version or higher because librpminspect
will now expect that functionality.

Also change the requirement on /usr/bin/annocheck to
'annobin-annocheck' so we are specifying package dependencies using
package names in the spec file.