Disable fedora-rawhide GitHub Action for now
There continues to be a gcc and kernel mismatch regarding the gcc
plugins provided by kernel-devel.  The gcc 11 builds in rawhide are
happening quickly and this keeps introducing mismatches with the
kernel.  We can't fix this, so just disabling the fedora-rawhide CI
job for now and will revisit it next week.