[new] Add 'xunit' output format support (#264)
librpminspect now supports a 3rd output format type:  XUnit

The RFE notes that XUnit and JUnit are the same, so I read this to get
the spec:

    https://llg.cubic.org/docs/junit/

I am not an expert on XUnit, so input welcome here.  One thing I
suppose it could do is show which tests the user skips by way of the
-E or -T command line options, but would that be "skipped" or
"disabled" in XUnit-speak?  Also looking for clarification that the
data is in the right place in the XUnit output.  The way the results
are collected in librpminspect are not easily mappable to what XUnit
defines, or at least I don't see an obvious way.  I am calling each
inspection a testcase and then each result from the inspection gets
reported as a <failure> or a <system-out>, but I do not know if this
is correct.  We'll try it this way and fix it up as necessary.

The output format reports the total number of tests and then the total
number of failures based on the severity threshold.  By default that
would be any result at VERIFY or higher.