[new] Create the 'badfuncs' inspection
The badfuncs inspection is the breakout of the forbidden IPv6 function
check that was in the elf inspection.  Instead of being restricted to
just IPv6 functions (well, by way of the reporting messages), it is
instead just a forbidden ELF symbol inspection.  If this does not
apply to your RPM, you can disable this function and still have the
elf inspection run.

NOTE: The elf inspection's include and exclude paths apply to this
function as well.