Fix spurious execstack failure
Commit 336d66d511fcbbd89092eeaab8e2ad904f4d29e2 introduced a
regression where the `get_elf_section` function would fail to traverse
ELF sections of ET_REL objects and thus fails the execstack check,
which looks for the GNU-stack note.

This fix continues to avoid the memcpy as the commit intended and
always traverses section headers to look for the requested section.