[lib] Add abspath() to canonicalize path strings
The abspath() function is similar to realpath(), but does not rely on
the filesystem.  All it does is read the actual path string and do
things like back up when it encounters a ".." and skip things like
".".  It is important to note that this function DOES NOT check to see
if the canonicalized path actually exists on the filesystem or not.
The purpose of this function is to reduce a complicated path with
.. redirects down to what it actually means.

This commit also adds some unit tests for the abspath() function.

The caller is responsible for freeing the returned string.

Signed-off-by: David Cantrell <dcantrell@redhat.com>