[new] Add the 'runpath' inspection to librpminspect
The runpath inspection checks ELF executables for invalid DT_RPATH or
DT_RUNPATH values.  The rules that control what is valid or invalid
are described in the data/generic.yaml configuration file template in
the 'runpath' section.  Basically, it comes down to this:

1) An ELF file can have DT_RPATH or DT_RUNPATH with allowed paths
mentioned.  Any relative path in the string is canonicalized down to
what it would actually refer to on the running system.  For example,
"/../../../usr/../usr/lib/" would canonicalize down to "/usr/lib" for
the purposes of checking it.

2) If a DT_RPATH or DT_RUNPATH value begins with $ORIGIN, the
following rules apply:
   a) "$ORIGIN" by itself is allowed.
   b) If $ORIGIN is followed by a string, then the value is advanced
   past "$ORIGIN" and the following happens:
      i) If a list of origin prefixes is defined in the configuration
      file, we check to see if any of those regexps match the prefix
      of the string.  If there is a match, we further advance the
      value past the match.
      ii) Once we have advanced past all allowed prefixes, we
      canonicalize the remaining part of the string and rule #1 is
      followed.

The inspection passes if all paths in the DT_RPATH and DT_RUNPATH
symbols are valid.  Invalid values are reported individually so you
know which part of the path string needs work.