[lib] Add strxmlescape() to strfuncs.c in librpminspect (#264)
Another string utility function.  This function takes a string and
allocates a new one that is a copy of the input string with the XML
special characters replaced with their special substrings for XML use.
The following special XML characters are replaced:

    < becomes &lt;
    > becomes &gt;
    ' becomes &apos;
    " becomes &quot;
    & becomes &amp;

This function is necessary to support outputting xunit from rpminspect
because we need to escape strings to use them as CDATA.

Signed-off-by: David Cantrell <dcantrell@redhat.com>