[fix] Output the system-out xunit portion as CDATA (#264)
The <system-out></system-out> portion of the xunit output lacked
correct CDATA syntax.  This commit aims to fix that, with the caveat
being I am not an XML expert.  Previously I was doing this for
reporting out results in xunit format:

    <system-out>%s</system-out>

Where the %s is the constructed string of the rpminspect results.  Now
the results are in this format:

    <system-out><![CDATA[%s]]></system-out>

Where the %s is the constructed string of the rpminspect results and
the constructed string has XML special characters escaped.

Signed-off-by: David Cantrell <dcantrell@redhat.com>