[fix] Fix block handling problems in the YAML config reader
This problem manifested in a strange way.  The test suite generates a
custom rpminspect.yaml file using the data/generic.yaml file as a
template.  In baseclass.py, it replaces some of the values with
runtime values that point to data locations for the test suite.  Then
the Python code writes out the YAML structure to a temporary file and
uses that.  The ordering of the blocks is not guaranteed, so the
resulting config file used by the test suite may be in a different
block ordering than the data/generic.yaml file.

What I found is that the read_cfgfile() function was not handling all
blocks correctly and some were only being read correctly when ordered
in a certain way in the config file.  So in trying to reproduce
problems using a real build and the rpminspect.yaml file from
rpminspect-data-fedora, it would work fine.  But recreating the same
test case in the test suite resulted in failure.

This patch refactors read_cfgfile() in a way that properly handles the
config file blocks and sections and do away with some of the more
complicated state tracking.  Rather than trying to treat the reads
based on the level you are at in the config file, just use the block
and group variables, even if it means there is slightly more code.
It's more readable and more obvious what is going on.

Signed-off-by: David Cantrell <dcantrell@redhat.com>