[ci] Disable Rust support in pip modules, more extra-ci fixes
Pass CRYPTOGRAPHY_DONT_BUILD_RUST=1 in the environment for PIP_CMD so
that Python modules installed with pip skip Rust parts.  Remove rustc
from the reqs.txt files that had that listed.

In the Ubuntu job, we have to use sudo because it's the only one not
running in a container.

In osdeps/debian/post.sh replace the pgrep/pkill with 'systemctl stop
clamav-freshclam.service'.

Add patchelf to the reqs.txt files for opensuse-leap and
opensuse-tumbleweed.

Signed-off-by: David Cantrell <dcantrell@redhat.com>